backend/requirements.txt
requests
flask
python-dotenv
flask-cors